Add additional functions and tune Monty's behaviour
Added a 'clear' as well as a 'confirm' function to enable the user to
clear the task list in a single command after asking for confirmation.
The delete command has also been modified to request a confirmation
beforehand.

Monty's response messages have been tweaked to add more personality to
the bot for a more fun and engaging user experience.

Some extra, unused pieces of code and some old JavaDoc comments have
been removed to make the code look cleaner.